crud.d_area: Add transformation functions
As geometries will only be stored as MultiPolygons, in case of
in- or outputting, single Polygons have to be transformed on the fly.

This commit adds functions to update the format of a Feature geometry dict
from Polygon to MultiPolygon and vice versa.

Relates to https://gitlab.gistools.geog.uni-heidelberg.de/giscience/heigit-disaster-portal/-/issues/6