Move to optional example data on startup
Previously a specific user and provider was always created if it
was not existing.
Added the environment variable 'CREATE_EXAMPLE_DATA_ON_STARTUP'.
It can be set to 'True' in the .env file to create a more general
example object for each collection if the collections are empty
(or only the admin user exists).