dap_api!: Improve password storage security
Switched from placeholder security (simple base64 encoding)
to more secure bcrypt_sha256 algorithm.
Added ENCRYPTION_SALT to environment variables.

The passed secret is used to extract the user from the DB
(resolving user by hashed_secret), as we can't use a function to filter
the rows of the Users table.

Therefore, we need a dedicated SALT to always produce the same hash,
instead of using random hashes + verify().
The second approach would be possible in a user-login + jwt token
scenario, where the user is provided next to a password.

BREAKIN_CHANGE: Existing hashed passwords need to be migrated manually.
This is tolerable as there are only a handfull of users.