api.db: increase postgis connection capacity
by default the sqlalchemy db connection allows for 5
simultaneous connections and 10 overflow connections.
This would result in deadlocks on the db for 50 requests
per second.

Postgres internally allows up to 100 connections by default.
With the new pool size of 20 and an overflow of 50
the portal should be able to handle a decent request amount.

Still requests should normally be sent in sequence, and not simultaneously.
Especially for long running requests with complex geometries.