backend.geoutil: Refactor bbox_bufer_percentage to buffer_bbox
the function previously buffered the bbox by the given percentage
of a square over the longer bbox side.
To keep the ratio, length and width are now buffered separately by
the given percentage.
The bbox can now also be buffered by a distance in coordinate units.
This can be combined with the percentage.