backend.geoutil: Refactor point_from_point_bearing_distance
the function previously would return coordinates with 14 decimal places.
Moreover, even if the point was generated on the x or y-axis,
the value on the axis that didn't change still changed:
e.g. for (3.555, 3.001) with bearing 90, the point returned could display
like ( x, 3.00099...) which is rather irritating.

The function was changed to round the returned coordinate values to
7 decimal places by default and keep the precision of the input value,
if the point is generated along the x or y-axis.

- added float_precision function to get the effective decimal places
  of a float value if it is below the defined precision limit.