build(docker): improve healthcheck & dependency
compose was already starting the api container when the db was
started but not ready, resulting in some unnecessary connection
failures.
It now properly waits until the DB is ready to accept connections.

Previously the DB container logs were spammed with FATAL errors:
role "root" does not exist.
This is fixed after doing the healthcheck with specified DB and user.